src_audio_and_fusion/pyproject.toml
[tool.poetry]
name = "10th-abaw"
version = "0.1.0"
description = ""
authors = ["Maxim Markitantov <m.markitantov@yandex.ru>"]
packages = [

]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
chimera-ml = { path = "../chimera_ml", develop = true }
torchaudio = "*"
torchvision = "*"
torchcodec = "*"
transformers = "*"
graphql-core = ">=3.1,<3.3"
jupyter = "^1.1.1"
opencv-python-headless = "^4.13.0.92"
funasr = "^1.3.1"
modelscope = "^1.34.0"
soundfile = "^0.13.1"
torchvggish = "^0.2"
ruamel-yaml = "^0.19.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."chimera_ml.plugins"]
10th_ABAW = "chimera_plugin:register"